.container_yolo26/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pillow==11.0.0
ultralytics==8.3.50
azure-storage-blob==12.24.0
PyJWT==2.10.1
